Moved function for building SOLPS transport files to its own module
